@@ -6,8 +6,8 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.1.1",
     "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^2.1.4"
   },
@@ -15,8 +15,8 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^14.0.4",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "gh-pages": "^3.2.3",
     "sass": "^1.50.0",
     "typescript": "^4.6.3"
@@ -43,8 +43,8 @@
     "deploy": "gh-pages -d build"
   },
   "engines": {
-    "node": "16.13.x",
-    "npm": "8.1.x"
+    "node": "16.14.x",
+    "npm": "8.5.x"
   },
   "eslintConfig": {
     "extends": [
