@@ -1,10 +1,14 @@
-import ReactDOM from 'react-dom';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-    <App />,
-    document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root')!);
+
+root.render(
+    <React.StrictMode>
+        <App />
+    </React.StrictMode>
 );
 
 reportWebVitals();
